datadog-agent/app/requirements.txt
Sanic==20.12.2
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
opentelemetry-instrumentation
